. modified the older version (Tex_IMAP) to use the new IMAP() function. . changed Tex_PutTextWithMovement() to just a hack. It substitutes '<+' and   '+>' with the s:PlaceHolder*() variables and then calls the new   Imap_PutText...
sorry for the short cvs log. I am running really late.

NOTE: This is a work in progress. the changed version of Tex_IMAP() is not
yet working quite robustly enought. It is sometimes generating more
arguments than are needed.

This commmit will be followed by another huge commit which is essentially
replacing all � characters with <++>, � with <+ and so on....

Therefore...

FROM HENCEFORTH:

Use <+ and +> instead of � and �

Better still, dont use any placeholders at all and use the new IMAP()
function.